Create questLog
The above doPost(e)-functions is meant to be set up as a google-script web-app, which listens to Habitica-webhooks on 'questInvited', 'questStarted' and 'questFinished' and 
writes the return data of the webhook to a google sheet. This sheet can be linked to a party description, so party members can look at the log file.   
What you need is:  
- Goolge sheet: enter in line 6
- The code above (deploy as web app) https://habitica.fandom.com/wiki/Webhook_processing_with_Google_Apps_Script#Google_Apps_Server_Script_.28Web-App.29  
- A webhook set up in Habitica (please use some tools and install a webhook that listens to 'questInvited', 'questStarted' and 'questFinished' (you might leave some out)) https://habitica.fandom.com/wiki/Webhook_processing_with_Google_Apps_Script#Third_party_tools: